Move dev-branch into a stable state.
Moved dev-branch into a stable state for any emergency deployments needed. Dev-branch has been unstable since the last two feature requests between a timespan of 60 minutes.
