+++ src/main/java/com/example/demo/domain/role/Role.java
@@ -2,6 +2,8 @@ package com.example.demo.domain.role;
 
 import com.example.demo.domain.authority.Authority;
 import lombok.*;
+import org.hibernate.annotations.OnDelete;
+import org.hibernate.annotations.OnDeleteAction;
 import org.hibernate.validator.constraints.UniqueElements;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
