+++ src/main/java/com/example/demo/domain/user/UserMapper.java
@@ -1,11 +1,8 @@
-package com.example.demo.domain.appUser;
+package com.example.demo.domain.user;
+
 import com.example.demo.domain.appUser.dto.PublicUserDTO;
 import org.mapstruct.Mapper;
 
-/**
- * Created by lazar on 3/2/2022.
- * Project name: demo
- **/
 @Mapper(componentModel = "spring")
 public interface UserMapper {
     PublicUserDTO convertUserToPublicUserDTO(User user);
